# openapi-schema.yaml
openapi: 3.0.0
info:
  title: NaviDocs API
  description: Comprehensive boat documentation management API with 16 supporting tables
  version: 1.0.0
  contact:
    name: NaviDocs Team
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.navidocs.com/v1
    description: Production server
paths:
...
  /boats/{boatId}/expenses:
    get:
      summary: List expenses for a boat
      operationId: listExpenses
      parameters:
        - name: boatId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
    post:
      summary: Create a new expense
      operationId: createExpense
      parameters:
        - name: boatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Expense created successfully
  /expenses/{expenseId}:
    get:
      summary: Get a specific expense
      operationId: getExpense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
    put:
      summary: Update an expense
      operationId: updateExpense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '200':
          description: Expense updated successfully
    delete:
      summary: Delete an expense
      operationId: deleteExpense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Expense deleted successfully
  /expenses/{expenseId}/approve:
    post:
      summary: Approve an expense
      operationId: approveExpense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Expense approved successfully

  # Warranties Endpoints
...
components:
  schemas:
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        boatId:
          type: integer
        amount:
          type: number
          format: decimal
        currency:
          type: string
        date:
          type: string
          format: date
        category:
          type: string
        receiptUrl:
          type: string
        ocrText:
          type: string
        splitUsers:
          type: object
        approvalStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateExpenseRequest:
      type: object
      required:
        - amount
        - date
      properties:
        amount:
          type: number
          format: decimal
        currency:
          type: string
        date:
          type: string
          format: date
        category:
          type: string
        receiptUrl:
          type: string
        ocrText:
          type: string
        splitUsers:
          type: object

    # Warranty Schema
...